Add some specific requirements

+++ src/main/java/kg/alatoo/sewing_industry_management/services/impl/DefectServiceImpl.java
@@ -1,10 +1,18 @@
-package kg.alatoo.sewing_industry_management.services;
+package kg.alatoo.sewing_industry_management.services.impl;
 
 import kg.alatoo.sewing_industry_management.dto.DefectDTO;
+import kg.alatoo.sewing_industry_management.dto.UserDTO;
 import kg.alatoo.sewing_industry_management.entities.Defect;
+import kg.alatoo.sewing_industry_management.entities.Product;
+import kg.alatoo.sewing_industry_management.entities.User;
+import kg.alatoo.sewing_industry_management.exception.DefectNotFoundException;
+import kg.alatoo.sewing_industry_management.exception.ProductNotFoundException;
+import kg.alatoo.sewing_industry_management.exception.UserNotFoundException;
 import kg.alatoo.sewing_industry_management.mappers.DefectMapper;
 import kg.alatoo.sewing_industry_management.repositories.DefectRepository;
+import kg.alatoo.sewing_industry_management.services.DefectService;
 import lombok.AllArgsConstructor;
+import org.springframework.data.jpa.repository.support.SimpleJpaRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,9 +36,10 @@ public class DefectServiceImpl implements DefectService {
     @Override
     public DefectDTO getDefectById(Long id) {
         Defect defect = defectRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Defect not found"));
+                .orElseThrow(() -> new UserNotFoundException("Defect with ID " + id + " not found"));
         return defectMapper.toDto(defect);
     }
+
 
     @Override
     public DefectDTO createDefect(DefectDTO defectDTO) {
@@ -42,7 +51,7 @@ public class DefectServiceImpl implements DefectService {
     @Override
     public DefectDTO updateDefect(Long id, DefectDTO defectDTO) {
         Defect existingDefect = defectRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Defect not found"));
+                .orElseThrow(() -> new DefectNotFoundException("Defect with ID " + id + " not found"));
 
         existingDefect.setDescription(defectDTO.getDescription());
         existingDefect.setQuantity(defectDTO.getQuantity());
@@ -53,6 +62,9 @@ public class DefectServiceImpl implements DefectService {
 
     @Override
     public void deleteDefect(Long id) {
+        if (!defectRepository.existsById(id)) {
+            throw new UserNotFoundException("Defect with ID " + id + " not found");
+        }
         defectRepository.deleteById(id);
     }
 }
