+++ src/main/java/kg/alatoo/sewing_industry_management/services/impl/RawMaterialServiceImpl.java
@@ -1,9 +1,12 @@
-package kg.alatoo.sewing_industry_management.services;
+package kg.alatoo.sewing_industry_management.services.impl;
 
 import kg.alatoo.sewing_industry_management.dto.RawMaterialDTO;
 import kg.alatoo.sewing_industry_management.entities.RawMaterial;
+import kg.alatoo.sewing_industry_management.exception.RawMaterialNotFoundException;
+import kg.alatoo.sewing_industry_management.exception.UserNotFoundException;
 import kg.alatoo.sewing_industry_management.mappers.RawMaterialMapper;
 import kg.alatoo.sewing_industry_management.repositories.RawMaterialRepository;
+import kg.alatoo.sewing_industry_management.services.RawMaterialService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -28,7 +31,7 @@ public class RawMaterialServiceImpl implements RawMaterialService {
     @Override
     public RawMaterialDTO getRawMaterialById(Long id) {
         RawMaterial rawMaterial = rawMaterialRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("RawMaterial not found"));
+                .orElseThrow(() -> new RawMaterialNotFoundException("Raw material with ID " + id + " not found"));
         return rawMaterialMapper.toDto(rawMaterial);
     }
 
@@ -42,12 +45,12 @@ public class RawMaterialServiceImpl implements RawMaterialService {
     @Override
     public RawMaterialDTO updateRawMaterial(Long id, RawMaterialDTO rawMaterialDTO) {
         RawMaterial existingRawMaterial = rawMaterialRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("RawMaterial not found"));
+                .orElseThrow(() -> new RawMaterialNotFoundException("Raw material with ID " + id + " not found"));
 
         existingRawMaterial.setName(rawMaterialDTO.getName());
         existingRawMaterial.setColor(rawMaterialDTO.getColor());
         existingRawMaterial.setQuantity(rawMaterialDTO.getQuantity());
-        existingRawMaterial.setDescription(rawMaterialDTO.getDescription());
+        existingRawMaterial.setStatus(rawMaterialDTO.getStatus());
 
         RawMaterial updatedRawMaterial = rawMaterialRepository.save(existingRawMaterial);
         return rawMaterialMapper.toDto(updatedRawMaterial);
@@ -55,7 +58,11 @@ public class RawMaterialServiceImpl implements RawMaterialService {
 
     @Override
     public void deleteRawMaterial(Long id) {
+        if (!rawMaterialRepository.existsById(id)) {
+            throw new RawMaterialNotFoundException("User with ID " + id + " not found");
+        }
         rawMaterialRepository.deleteById(id);
+
     }
 }
